rag/models/llava_model.py
"""
LLaVA Model Wrapper

Handles LLaVA model loading and inference for multi-modal reasoning.
"""

from typing import Optional, List, Union
import torch
from PIL import Image
from loguru import logger


class LLaVAModel:
    """LLaVA model wrapper for vision-language reasoning"""
    
    def __init__(self, model_name: str, device: str = "auto", quantization: Optional[str] = None):
        """
        Initialize LLaVA model.
        
        Args:
            model_name: HuggingFace model identifier
            device: Device to load model on
            quantization: Quantization configuration
        """
        self.model_name = model_name
        self.device = device
        self.quantization = quantization
        self.model = None
        self.processor = None
        self.tokenizer = None
        
        # TODO: Load model components
        self._load_model()
        
    def _load_model(self):
        """Load LLaVA model components"""
        try:
            logger.info(f"Loading LLaVA model: {self.model_name}")
            
            from transformers import LlavaNextProcessor, LlavaNextForConditionalGeneration
            
            # Load processor and model
            self.processor = LlavaNextProcessor.from_pretrained(self.model_name)
            
            # Configure model loading based on quantization
            model_kwargs = {
                "torch_dtype": torch.float16,
                "device_map": self.device if self.device != "auto" else "auto",
            }
            
            if self.quantization == "4bit":
                from transformers import BitsAndBytesConfig
                model_kwargs["quantization_config"] = BitsAndBytesConfig(
                    load_in_4bit=True,
                    bnb_4bit_compute_dtype=torch.float16
                )
            elif self.quantization == "8bit":
                model_kwargs["load_in_8bit"] = True
            
            self.model = LlavaNextForConditionalGeneration.from_pretrained(
                self.model_name,
                **model_kwargs
            )
            
            logger.info("LLaVA model loaded successfully")
            
        except Exception as e:
            logger.error(f"Failed to load LLaVA model: {e}")
            raise
            
    def generate_response(
        self, 
        prompt: str, 
        image: Optional[Union[Image.Image, str]] = None,
        max_tokens: int = 512,
        temperature: float = 0.7
    ) -> str:
        """
        Generate response using LLaVA model.
        
        Args:
            prompt: Text prompt/question
            image: PIL Image or path to image file
            max_tokens: Maximum tokens to generate
            temperature: Sampling temperature
            
        Returns:
            Generated response text
        """
        try:
            logger.info("Generating LLaVA response")
            
            if not self.is_loaded():
                return "Error: LLaVA model not loaded"
            
            # Handle image input
            if image is not None:
                if isinstance(image, str):
                    image = Image.open(image)
                elif isinstance(image, bytes):
                    import io
                    image = Image.open(io.BytesIO(image))
            else:
                logger.warning("No image provided for LLaVA inference")
                return "Error: LLaVA requires an image input"
            
            # Prepare inputs
            inputs = self.processor(prompt, image, return_tensors="pt")
            
            # Move inputs to device if needed
            if hasattr(self.model, 'device'):
                inputs = {k: v.to(self.model.device) if isinstance(v, torch.Tensor) else v 
                         for k, v in inputs.items()}
            
            # Generate response
            with torch.no_grad():
                generate_kwargs = {
                    "max_new_tokens": max_tokens,
                    "do_sample": temperature > 0,
                    "pad_token_id": self.processor.tokenizer.eos_token_id,
                }
                
                if temperature > 0:
                    generate_kwargs.update({
                        "temperature": temperature,
                        "top_p": 0.9,
                    })
                
                outputs = self.model.generate(
                    **inputs,
                    **generate_kwargs
                )
            
            # Decode response
            generated_tokens = outputs[0][inputs["input_ids"].shape[1]:]
            response = self.processor.tokenizer.decode(generated_tokens, skip_special_tokens=True)
            
            logger.info(f"Generated LLaVA response: {response[:100]}...")
            return response.strip()
            
        except Exception as e:
            logger.error(f"LLaVA generation failed: {e}")
            return f"Error: {e}"
            
    def is_loaded(self) -> bool:
        """Check if model is loaded"""
        return self.model is not None
        
    def unload(self):
        """Unload model from memory"""
        if self.model:
            del self.model
            del self.processor
            del self.tokenizer
            torch.cuda.empty_cache()
            
        self.model = None
        self.processor = None
        self.tokenizer = None
        
        logger.info("LLaVA model unloaded")
        
    def get_memory_usage(self) -> dict:
        """Get model memory usage"""
        # TODO: Implement memory tracking
        return {
            "model_loaded": self.is_loaded(),
            "estimated_memory_mb": 14000 if self.is_loaded() else 0
        }