requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
scikit-learn>=1.1.0
nltk>=3.7
spacy>=3.4.0
pathlib
typing-extensions
opencv-python>=4.5.0
Pillow>=8.0.0
loguru>=0.6.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
llava-torch>=1.1.0
google-genai>=0.3.0 
